+++ src/toast_notification.py
@@ -1,44 +1,10 @@
 import math
 import os
-from enum import Enum
 from qtpy.QtGui import QGuiApplication
 from qtpy.QtCore import Qt, QPropertyAnimation, QPoint, QTimer, QSize, QMargins, QRect, Signal
 from qtpy.QtGui import QPixmap, QIcon, QColor, QFont, QImage, qRgba, QFontMetrics
 from qtpy.QtWidgets import QDialog, QPushButton, QLabel, QGraphicsOpacityEffect, QWidget
-
-
-class ToastPreset(Enum):
-    SUCCESS = 1
-    WARNING = 2
-    ERROR = 3
-    INFORMATION = 4
-    SUCCESS_DARK = 5
-    WARNING_DARK = 6
-    ERROR_DARK = 7
-    INFORMATION_DARK = 8
-
-
-class ToastIcon(Enum):
-    SUCCESS = 1
-    WARNING = 2
-    ERROR = 3
-    INFORMATION = 4
-    CLOSE = 5
-
-
-class ToastPosition(Enum):
-    BOTTOM_RIGHT = 1
-    BOTTOM_LEFT = 2
-    BOTTOM_MIDDLE = 3
-    TOP_RIGHT = 4
-    TOP_LEFT = 5
-    TOP_MIDDLE = 6
-
-
-class ToastButtonAlignment(Enum):
-    TOP = 1
-    MIDDLE = 2
-    BOTTOM = 3
+from .toast_enums import ToastPreset, ToastIcon, ToastPosition, ToastButtonAlignment
 
 
 class ToastNotification(QDialog):
